feat(rabbitmq): 처리 완료 데이터 RabbitMQ 발행 기능 구현
정제 및 검증이 완료된 센서 데이터를 후속 서비스(Rule Engine 등)로 전달하기 위해 RabbitMQ 발행 파이프라인을 구축함.

- DirectExchange 및 Jackson2JsonMessageConverter 기반 RabbitTemplate 설정 (RabbitMQConfig)
- JSON 직렬화 품질 향상을 위해 ObjectMapper 설정 추가 (AppConfig)
- ParsedSensorMessage 객체를 지정된 Exchange/RoutingKey로 발행하는 RabbitMqPublisher 구현

+++ src/main/java/com/nhnacademy/processing/dto/api/SensorContext.java
@@ -2,7 +2,7 @@ package com.nhnacademy.processing.dto.api;
 
 public record SensorContext(
         String devEui,
-        Long roomId,
-        Long teamId
+        Integer roomId,
+        Integer teamId
 ) {
 }
